+++ armada_backend/hermes_init.py
@@ -1,29 +1,19 @@
-from __future__ import print_function
-
 import json
 import socket
-import sys
 import time
 
 import requests
 
-from armada_backend.utils import get_container_ssh_address
-from armada_command.consul.consul import consul_query, ConsulException
+from armada_backend.utils import get_container_ssh_address, get_logger
+from armada_command.consul.consul import consul_query
 
 HERMES_DIRECTORY = '/etc/opt'
 
 
-def _print_err(*objs):
-    print(*objs, file=sys.stderr)
-
-
 def _consul_discover(service_name):
     service_addresses = set()
-    try:
-        query = 'health/service/{service_name}'.format(service_name=service_name)
-        instances = consul_query(query)
-    except ConsulException:
-        pass
+    query = 'health/service/{service_name}'.format(service_name=service_name)
+    instances = consul_query(query)
 
     for instance in instances:
         service_checks_statuses = (check['Status'] for check in instance['Checks'])
@@ -53,7 +43,7 @@ def _wait_for_armada_start():
         except:
             pass
     if not armada_is_running:
-        _print_err('Could not connect to armada.')
+        get_logger().error('Could not connect to armada.')
         return
 
 
@@ -62,17 +52,22 @@ def _get_courier_addresses():
     courier_is_running = False
 
     timeout_expiration = time.time() + 30
+    last_exception = None
     while time.time() < timeout_expiration:
         time.sleep(1)
         try:
             courier_addresses = _consul_discover('courier')
+            last_exception = None
             if courier_addresses:
                 courier_is_running = True
                 break
-        except:
-            pass
-    if not courier_is_running:
-        print('No running couriers found.')
+        except Exception as e:
+            last_exception = e
+    if last_exception is not None:
+        get_logger().error('Could not determine if courier is running:')
+        get_logger().exception(last_exception)
+    elif not courier_is_running:
+        get_logger().info('No running couriers found.')
     return courier_addresses
 
 
@@ -84,10 +79,11 @@ def _fetch_hermes_from_couriers(courier_addresses):
             payload = {'ssh': my_ssh_address, 'path': HERMES_DIRECTORY}
             requests.post(courier_url, json.dumps(payload))
         except Exception as e:
-            _print_err('Fetching all sources from courier {courier_address} failed: {e}'.format(**locals()))
+            get_logger().error('Fetching all sources from courier {courier_address} failed:'.format(**locals()))
+            get_logger().exception(e)
 
 
-if __name__ == '__main__':
+def main():
     _wait_for_armada_start()
 
     # We fetch data from courier as soon as possible to cover most common case of 1 courier running.
@@ -98,3 +94,7 @@ if __name__ == '__main__':
     time.sleep(11)
     new_courier_addresses = _get_courier_addresses() - courier_addresses
     _fetch_hermes_from_couriers(new_courier_addresses)
+
+
+if __name__ == '__main__':
+    main()