+++ armada_backend/recover_saved_containers.py
@@ -1,15 +1,13 @@
-from __future__ import print_function
 import argparse
-from collections import Counter
-from time import sleep
 import json
 import os
-import traceback
 import sys
+import traceback
+from collections import Counter
+from time import sleep
 
-from armada_backend.api_run import print_err
 from armada_backend.api_ship import wait_for_consul_ready
-from armada_backend.utils import get_container_parameters, get_local_containers_ids
+from armada_backend.utils import get_container_parameters, get_local_containers_ids, get_logger
 from armada_command import armada_api
 
 RECOVERY_COMPLETED_PATH = '/tmp/recovery_completed'
@@ -34,9 +32,9 @@ def _get_local_running_containers():
 
 
 def _recover_container(container_parameters):
-    print_err('Recovering: {}...\n'.format(json.dumps(container_parameters)))
+    get_logger().info('Recovering: {}...\n'.format(json.dumps(container_parameters)))
     recovery_result = armada_api.post('run', container_parameters)
-    print_err('Recovered container: {}'.format(json.dumps(recovery_result)))
+    get_logger().info('Recovered container: {}'.format(json.dumps(recovery_result)))
 
 
 def _multiset_difference(a, b):
@@ -52,7 +50,7 @@ def recover_saved_containers(saved_containers):
     containers_to_be_recovered = _multiset_difference(saved_containers, running_containers)
     recovery_retry_count = 0
     while containers_to_be_recovered and recovery_retry_count < RECOVERY_RETRY_LIMIT:
-        print_err("Recovering containers: {}".format(json.dumps(containers_to_be_recovered)))
+        get_logger().info("Recovering containers: {}".format(json.dumps(containers_to_be_recovered)))
         for container_parameters in containers_to_be_recovered:
             _recover_container(container_parameters)
         sleep(DELAY_BETWEEN_RECOVER_RETRY_SECONDS)
@@ -67,23 +65,23 @@ def _recover_saved_containers_from_path(saved_containers_path):
         saved_containers = _load_saved_containers_parameters_list(saved_containers_path)
         not_recovered = recover_saved_containers(saved_containers)
         if not_recovered:
-            print_err('Following containers were not recovered: ', not_recovered)
+            get_logger().error('Following containers were not recovered: ', not_recovered)
             return False
         else:
             return True
     except:
         traceback.print_exc()
-        print_err('Unable to recover from {saved_containers_path}.'.format(**locals()))
+        get_logger().error('Unable to recover from {saved_containers_path}.'.format(**locals()))
     return False
 
 
 def _check_if_we_should_recover(saved_containers_path):
     try:
         if int(os.environ.get('DOCKER_START_TIMESTAMP')) > int(os.path.getmtime(saved_containers_path)):
-            print_err('Docker daemon restart detected.')
+            get_logger().info('Docker daemon restart detected.')
             return True
         else:
-            print_err('No need to recover.')
+            get_logger().info('No need to recover.')
             return False
     except:
         return False