+++ armada_backend/run_consul.py
@@ -4,6 +4,7 @@ import random
 import shutil
 
 import consul_config
+from utils import get_logger
 
 
 def _get_runtime_settings():
@@ -27,10 +28,10 @@ def _get_runtime_settings():
 
     ship_ips = runtime_settings.get('ships', [])
     consul_mode = consul_config.ConsulMode.BOOTSTRAP
-    if runtime_settings.get('is_commander') == True:
+    if runtime_settings.get('is_commander') is True:
         if ship_ips and len(ship_ips) > 0:
             consul_mode = consul_config.ConsulMode.SERVER
-    if runtime_settings.get('is_commander') == False:
+    if runtime_settings.get('is_commander') is False:
         consul_mode = consul_config.ConsulMode.CLIENT
 
     if runtime_settings.get('datacenter'):
@@ -44,13 +45,14 @@ def _get_runtime_settings():
 def main():
     ship_name, consul_mode, ship_ips, datacenter = _get_runtime_settings()
     ship_external_ip = os.environ.get('SHIP_EXTERNAL_IP', '')
-    consul_config_content = consul_config.get_consul_config(**locals())
+    consul_config_content = consul_config.get_consul_config(consul_mode, ship_ips, datacenter, ship_name,
+                                                            ship_external_ip)
 
     with open(consul_config.CONFIG_PATH, 'w') as config_file:
         config_file.write(consul_config_content)
 
     command = '/usr/local/bin/consul agent -config-file {config_path}'.format(config_path=consul_config.CONFIG_PATH)
-    print 'RUNNING: ' + command
+    get_logger().info('RUNNING: {}'.format(command))
 
     args = command.split()
     os.execv(args[0], args)