+++ armada_backend/save_running_containers.py
@@ -5,10 +5,9 @@ import shutil
 import sys
 import traceback
 
-from api_run import print_err
 from armada_backend.api_ship import wait_for_consul_ready
 from armada_backend.recover_saved_containers import RECOVERY_COMPLETED_PATH
-from armada_backend.utils import get_container_parameters, get_ship_name
+from armada_backend.utils import get_container_parameters, get_ship_name, get_logger
 from armada_backend.utils import get_local_containers_ids
 from armada_command.consul import kv
 
@@ -47,7 +46,7 @@ def _is_recovery_completed():
 def main():
     args = _parse_args()
     if not args.force and not _is_recovery_completed():
-        print('Recovery is not completed. Aborting saving running containers.')
+        get_logger().warning('Recovery is not completed. Aborting saving running containers.')
         return
     saved_containers_path = args.saved_containers_path
     try:
@@ -62,20 +61,20 @@ def main():
                     containers_parameters_list.append(container_parameters)
             except:
                 errors_count += 1
-                print_err('ERROR on getting container parameters for {container_id}:'.format(**locals()))
+                get_logger().error('ERROR on getting container parameters for {container_id}:'.format(**locals()))
                 traceback.print_exc()
         containers_parameters_list.sort()
         # Don't overwrite saved containers' list if it would become empty because of errors.
         if containers_parameters_list or not errors_count:
             _save_containers_parameters_list_in_file(containers_parameters_list, saved_containers_path)
-            print('Containers have been saved to {saved_containers_path}.'.format(**locals()))
+            get_logger().info('Containers have been saved to {saved_containers_path}.'.format(**locals()))
             try:
                 _save_containers_parameters_list_in_kv_store(containers_parameters_list)
-                print('Containers have been saved to kv store.')
+                get_logger().info('Containers have been saved to kv store.')
             except:
                 traceback.print_exc()
         else:
-            print('Aborted saving container because of errors.')
+            get_logger().info('Aborted saving container because of errors.')
     except:
         traceback.print_exc()
         sys.exit(1)