+++ docker-containers/microservice/scripts/install_packages.sh
@@ -3,9 +3,9 @@
 apt-get update
 apt-get upgrade -y
 apt-get install -y supervisor python python-dev python-pip curl mc software-properties-common wget
-add-apt-repository -y ppa:vbernat/haproxy-1.5
+add-apt-repository -y ppa:vbernat/haproxy-1.6
 apt-get update
 apt-get install -y haproxy
 
 pip install -U pip
-pip install -U docker-py web.py requests
+pip install -U docker-py web.py "requests==2.9.1"